+++ run/one_click.sh
@@ -1,3 +1,3 @@
-sh run/compare_glb.sh
-sh run/static_optm_test.sh
-sh run/compare_dyn.sh
+sh run/compare_glb.sh 
+sh run/static_optm_test.sh 300 20 360
+sh run/compare_dyn.sh 300 20 360